Add cosmic-text as an option for the text renderer
Cosmic-text is still changing somewhat rapidly, but should provide a way
to handle RTL/bidir/etc text with proper shaping. So I thought I'd try
adding support for it here.

This isn't quite working right when I try it with Arabic text (maybe
partly that it needs alpha blending when compositing glyphs?). And it
needs to request the right font.